Moved addNoise to env. Detecting collision now considers open/close.
